moving files into test folder following PT structure
